backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-neo4j==0.1.0
neo4j-graphrag>=0.1.0
python-dotenv==1.0.0
neo4j>=5.25.0
# HTTP client used to call the external embeddings service
requests==2.32.5
